Upgrade to Feathers 5 and latest dependencies
