--- .babelrc
@@ -1,10 +1,0 @@
-{
-  "plugins": [ "add-module-exports", "transform-export-extensions" ],
-  "presets": [
-    ["env", {
-      "targets": {
-        "node": "7.0"
-      }
-    }]
-  ]
-}
