+++ .gitignore
@@ -31,4 +31,4 @@ node_modules
 .lock-wscript
 
 # The compiled/babelified modules
-lib/
+data/
