+++ lib/index.js
@@ -1,7 +1,8 @@
-import database from './database';
-import collection from './collection';
-import user from './user';
 import makeDebug from 'debug';
+
+import database from './database.js';
+import collection from './collection.js';
+import user from './user.js';
 
 const debug = makeDebug('feathers-mongodb-management');
