+++ package.json
@@ -3,7 +3,8 @@
   "description": "Feathers service adapters for managing MongoDB databases, collections and users",
   "version": "1.0.0",
   "homepage": "https://github.com/feathersjs-ecosystem/feathers-mongodb-management",
-  "main": "lib/",
+  "main": "lib/index.js",
+  "type": "module",
   "keywords": [
     "feathers",
     "feathers-plugin"
@@ -22,25 +23,22 @@
     "url": "https://github.com/feathersjs-ecosystem/feathers-mongodb-management/issues"
   },
   "engines": {
-    "node": ">= 8.0.0"
+    "node": ">= 14.0.0"
   },
   "scripts": {
-    "prepublish": "npm run compile",
     "publish": "git push origin --tags && npm run changelog && git push origin",
     "release:patch": "npm version patch && npm publish",
     "release:minor": "npm version minor && npm publish",
     "release:major": "npm version major && npm publish",
     "changelog": "github_changelog_generator -u feathersjs-ecosystem -p feathers-mongodb-management && git add CHANGELOG.md && git commit -am \"Updating changelog\"",
-    "compile": "shx rm -rf lib/ && babel -d lib/ --copy-files src/",
-    "watch": "babel --watch -d lib/ --copy-files src/",
-    "lint": "semistandard src/**/*.js test/**/*.js --fix",
-    "mocha": "cross-env NODE_CONFIG_DIR=./test/config/ mocha --exit --require babel-core/register",
+    "lint": "semistandard lib/**/*.js test/**/*.js --fix",
+    "mocha": "cross-env NODE_CONFIG_DIR=./test/config/ mocha test/*.test.js --exit",
     "coverage": "nyc npm run mocha",
-    "test": "npm run compile && npm run lint && npm run coverage",
-    "start": "npm run compile && node example/app"
+    "mongodb": "run-rs -v 4.0.0",
+    "test": "npm run lint && npm run coverage",
+    "start": "node example/app"
   },
   "semistandard": {
-    "parser": "babel-eslint",
     "sourceType": "module",
     "env": [
       "mocha"
@@ -50,30 +48,25 @@
     "lib": "lib"
   },
   "dependencies": {
-    "debug": "^4.0.0",
-    "@feathersjs/feathers": "^3.1.7",
-    "@feathersjs/commons": "^4.0.0",
-    "@feathersjs/errors": "^3.3.0",
-    "@feathersjs/adapter-commons": "^2.0.0",
-    "mongodb": "^3.6.2",
-    "sift": "^8.0.0"
+    "@feathersjs/adapter-commons": "^5.0.0-pre.15",
+    "@feathersjs/commons": "^5.0.0-pre.15",
+    "@feathersjs/errors": "^5.0.0-pre.15",
+    "@feathersjs/feathers": "^5.0.0-pre.15",
+    "debug": "^4.3.3",
+    "mongodb": "^4.2.1",
+    "sift": "^15.1.0"
   },
   "devDependencies": {
-    "@feathersjs/configuration": "^1.0.2",
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^10.0.3",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.7.0",
-    "chai": "^4.2.0",
+    "@feathersjs/configuration": "^5.0.0-pre.15",
+    "chai": "^4.3.4",
     "chai-lint": "^0.1.1",
     "containerized": "^1.0.2",
-    "cross-env": "^5.2.0",
-    "eslint": "^6.8.0",
-    "mocha": "^6.2.0",
-    "nyc": "^15.0.1",
-    "semistandard": "^14.0.0",
-    "shx": "^0.3.2"
+    "cross-env": "^7.0.3",
+    "eslint": "^8.4.1",
+    "mocha": "^9.1.3",
+    "nyc": "^15.1.0",
+    "run-rs": "^0.7.5",
+    "semistandard": "^16.0.1",
+    "shx": "^0.3.3"
   }
 }
